Factor out the new table list component to a new Vocab specific component
